+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `v1.GetTimeOffsMapped()` function
+
 ## [0.3.0] - 2023-07-05
 
 ### Added